refactor(types,clerk-js): Extract SignUpFuture typed params into types
