+++ packages/clerk-js/src/core/resources/SignUp.ts
@@ -14,11 +14,15 @@ import type {
   PreparePhoneNumberVerificationParams,
   PrepareVerificationParams,
   PrepareWeb3WalletVerificationParams,
-  SetActiveNavigate,
   SignUpAuthenticateWithWeb3Params,
   SignUpCreateParams,
   SignUpField,
+  SignUpFutureCreateParams,
+  SignUpFutureEmailCodeVerifyParams,
+  SignUpFutureFinalizeParams,
+  SignUpFuturePasswordParams,
   SignUpFutureResource,
+  SignUpFutureSSoParams,
   SignUpIdentificationField,
   SignUpJSON,
   SignUpJSONSnapshot,
@@ -524,7 +528,8 @@ class SignUpFuture implements SignUpFutureResource {
     return { captchaToken, captchaWidgetType, captchaError };
   }
 
-  async create({ transfer }: { transfer?: boolean }): Promise<{ error: unknown }> {
+  async create(params: SignUpFutureCreateParams): Promise<{ error: unknown }> {
+    const { transfer } = params;
     return runAsyncResourceTask(this.resource, async () => {
       const { captchaToken, captchaWidgetType, captchaError } = await this.getCaptchaToken();
       await this.resource.__internal_basePost({
@@ -534,7 +539,8 @@ class SignUpFuture implements SignUpFutureResource {
     });
   }
 
-  async password({ emailAddress, password }: { emailAddress: string; password: string }): Promise<{ error: unknown }> {
+  async password(params: SignUpFuturePasswordParams): Promise<{ error: unknown }> {
+    const { emailAddress, password } = params;
     return runAsyncResourceTask(this.resource, async () => {
       const { captchaToken, captchaWidgetType, captchaError } = await this.getCaptchaToken();
 
@@ -561,7 +567,8 @@ class SignUpFuture implements SignUpFutureResource {
     });
   }
 
-  async verifyEmailCode({ code }: { code: string }): Promise<{ error: unknown }> {
+  async verifyEmailCode(params: SignUpFutureEmailCodeVerifyParams): Promise<{ error: unknown }> {
+    const { code } = params;
     return runAsyncResourceTask(this.resource, async () => {
       await this.resource.__internal_basePost({
         body: { strategy: 'email_code', code },
@@ -570,15 +577,8 @@ class SignUpFuture implements SignUpFutureResource {
     });
   }
 
-  async sso({
-    strategy,
-    redirectUrl,
-    redirectUrlComplete,
-  }: {
-    strategy: string;
-    redirectUrl: string;
-    redirectUrlComplete: string;
-  }): Promise<{ error: unknown }> {
+  async sso(params: SignUpFutureSSoParams): Promise<{ error: unknown }> {
+    const { strategy, redirectUrl, redirectUrlComplete } = params;
     return runAsyncResourceTask(this.resource, async () => {
       const { captchaToken, captchaWidgetType, captchaError } = await this.getCaptchaToken();
       await this.resource.__internal_basePost({
@@ -603,7 +603,8 @@ class SignUpFuture implements SignUpFutureResource {
     });
   }
 
-  async finalize({ navigate }: { navigate?: SetActiveNavigate } = {}): Promise<{ error: unknown }> {
+  async finalize(params?: SignUpFutureFinalizeParams): Promise<{ error: unknown }> {
+    const { navigate } = params || {};
     return runAsyncResourceTask(this.resource, async () => {
       if (!this.resource.createdSessionId) {
         throw new Error('Cannot finalize sign-up without a created session.');
